pyproject.toml
[project]
name = "airp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "google-generativeai>=0.8.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.0",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "I"]  # include I = import sorting

[tool.ruff.lint.isort]
# optional custom import sorting settings
combine-as-imports = true
known-third-party = ["fastapi", "pandas"]
